refactor: MultiTagInput and DateInputRefactor
